+++ src/Resources/contao/languages/de/tl_settings.php
@@ -1,5 +1,11 @@
 <?php 
 
+// Abschnitt Globale Einstellungen
+$GLOBALS['TL_LANG']['tl_settings']['characterSet'] = array('Zeichensatz', 'Damit Sonderzeichen richtig dargestellt werden, wird die Verwendung von UTF-8 empfohlen.');
+
+// Abschnitt Dateien und Bilder
+$GLOBALS['TL_LANG']['tl_settings']['editableFiles'] = array('Editierbare Dateien', 'Hier können Sie eine kommagetrennte Liste von Dateitypen eingeben, die mit dem Quelltexteditor bearbeitet werden dürfen.');
+
 // Abschnitt Speicherzeiten
 $GLOBALS['TL_LANG']['tl_settings']['timeout_legend'] = 'Speicherzeiten';
 $GLOBALS['TL_LANG']['tl_settings']['undoPeriod'] = array('Speicherzeit für Undo-Schritte', 'Hier können Sie die Speicherzeit für Undo-Schritte in Sekunden eingeben (24 Stunden = 86400 Sekunden).');