+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Einstellungen
 
+## Version 1.0.0 (2016-06-16)
+
+* Speicherzeiten an alte Stelle verschoben
+* Zeichensatz hinzugefügt
+* Editierbare Dateien hinzugefügt
+
 ## Version 0.0.2 (2016-06-16)
 
 * Speicherzeiten hinzugefügt